fix: corrigido erro que mostrava todos os torneios na tela Meus Torneios e refatorado o componente TournamentDetails
