+++ src/views/Tournament/TournamentDetails/TournamentDetails.tsx
@@ -1,10 +1,12 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { useEffect, useState } from "react";
+/* eslint-disable react-hooks/exhaustive-deps */
+import { useEffect } from "react";
 import { Link, useParams } from "react-router-dom";
-import api from "@/services/api";
 import { useTournament } from "@/hooks/useTournament";
-import { useFlash } from "@/hooks/useFlash";
-import type { Category, Group, Match } from "@/types/types";
+import { useTournamentAuth } from "@/hooks/useTournamentAuth";
+import { useTournamentOwnership } from "@/hooks/useTournamentOwnership";
+import { useTournamentCategories } from "@/hooks/useTournamentCategories";
+import { useTournamentScores } from "@/hooks/useTournamentScores";
+import { useTournamentActions } from "@/hooks/useTournamentActions";
 import {
   CategoryList,
   CategoryManager,
@@ -16,37 +18,39 @@ import { Modal } from "@/components";
 const TournamentDetails = () => {
   const { id } = useParams<{ id: string }>();
   const { tournament, fetchTournamentById, loading } = useTournament();
-  const { showFlash } = useFlash();
+  const { isLoggedIn, userId } = useTournamentAuth();
+  const { isOwner, canEdit } = useTournamentOwnership(
+    tournament,
+    userId,
+    isLoggedIn
+  );
 
-  const [scores, setScores] = useState<
-    Record<string, { team1: string; team2: string }>
-  >({});
-  const [originalScores, setOriginalScores] = useState<
-    Record<string, { team1: string; team2: string }>
-  >({});
-  const [isOwner, setIsOwner] = useState<boolean>(false);
-  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
-  const [userId, setUserId] = useState<string | null>(null);
-  const [isEditMode, setIsEditMode] = useState<boolean>(false);
-  const [isEditingInfo, setIsEditingInfo] = useState<boolean>(false);
-  const [categories, setCategories] = useState<Category[]>([]);
-  const [originalCategories, setOriginalCategories] = useState<Category[]>([]);
+  const {
+    categories,
+    setCategories,
+    resetCategories,
+    updateOriginalCategories,
+  } = useTournamentCategories(tournament);
 
-  // Permissão para editar (precisa estar logado E ser o dono)
-  const canEdit = isLoggedIn && isOwner;
+  const { scores, originalScores, setScores, setOriginalScores } =
+    useTournamentScores(tournament);
 
-  /* Verifica autenticação ao montar */
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    const storedUserId = localStorage.getItem("userId");
-
-    if (token && storedUserId) {
-      setIsLoggedIn(true);
-      setUserId(storedUserId);
-    } else {
-      setIsLoggedIn(false);
-    }
-  }, []);
+  const {
+    isEditMode,
+    isEditingInfo,
+    setIsEditMode,
+    setIsEditingInfo,
+    handleCleanCategories,
+    handleCancelEdit,
+    handleSaveTournamentInfo,
+    handleSaveCategories,
+    handleDrawCategory,
+    handleScoreUpdate,
+  } = useTournamentActions(
+    tournament?._id,
+    fetchTournamentById,
+    updateOriginalCategories
+  );
 
   /* Buscar torneio ao montar */
   useEffect(() => {
@@ -54,258 +58,32 @@ const TournamentDetails = () => {
     fetchTournamentById(id);
   }, [id]);
 
-  /* Verifica se o usuário é o dono do torneio */
-  useEffect(() => {
-    if (tournament && userId) {
-      // Verifica se tournament.user é string (ID) ou objeto (populado)
-      const tournamentUserId =
-        typeof tournament.user === "string"
-          ? tournament.user
-          : tournament.user?._id || tournament.user?.id;
-
-      const owner = tournamentUserId === userId;
-      setIsOwner(owner);
-    } else {
-      setIsOwner(false);
-    }
-  }, [tournament, userId]);
-
-  // Carrega categorias do torneio
-  useEffect(() => {
-    if (tournament?.categories) {
-      setCategories(tournament.categories);
-      setOriginalCategories(JSON.parse(JSON.stringify(tournament.categories)));
-    }
-  }, [tournament]);
-
-  // Inicializa os placares (estado atual e original)
-  useEffect(() => {
-    if (!tournament?.categories?.length) return;
-
-    const init: Record<string, { team1: string; team2: string }> = {};
-
-    tournament.categories.forEach((category: Category, catIndex: number) => {
-      category.groups?.forEach((group: Group, gi: number) => {
-        group.matches.forEach((match: Match, mi: number) => {
-          if (match.score) {
-            const [a, b] = match.score.split("-").map((v: string) => v.trim());
-            init[`${catIndex}-${gi}-${mi}`] = {
-              team1: a || "",
-              team2: b || "",
-            };
-          } else {
-            init[`${catIndex}-${gi}-${mi}`] = { team1: "", team2: "" };
-          }
-        });
-      });
-    });
-
-    setScores(init);
-    setOriginalScores(init);
-  }, [tournament]);
-
-  // Função para converter data para ISO
-  const dateToISO = (dateString: string): string => {
-    if (!dateString) return "";
-    const [year, month, day] = dateString.split("-");
-    return new Date(
-      parseInt(year),
-      parseInt(month) - 1,
-      parseInt(day),
-      12,
-      0,
-      0
-    ).toISOString();
-  };
-
-  // Limpar alterações não salvas (restaura do backup)
-  const handleCleanCategories = () => {
-    const confirmed = window.confirm(
-      "Tem certeza que deseja limpar todas as alterações não salvas?"
+  // Estado de carregamento ou erro
+  if (loading) {
+    return (
+      <section>
+        <div className="flex flex-col w-full min-h-screen items-center justify-center pt-22 px-5 bg-white">
+          <p className="text-gray-600">Carregando torneio...</p>
+        </div>
+      </section>
     );
+  }
 
-    if (confirmed) {
-      setCategories(JSON.parse(JSON.stringify(originalCategories))); // Restaura do backup
-      showFlash("Alterações limpas!", "info");
-    }
-  };
-
-  // Cancelar edição (limpa e volta para visualização)
-  const handleCancelEdit = async () => {
-    const confirmed = window.confirm(
-      "Tem certeza que deseja cancelar? Todas as alterações não salvas serão perdidas."
+  if (!tournament) {
+    return (
+      <section>
+        <div className="flex flex-col w-full min-h-screen items-center justify-center pt-22 px-5 bg-white">
+          <p className="text-gray-600 mb-4">Torneio não encontrado!</p>
+          <Link
+            to={isLoggedIn ? "/my-tournaments" : "/"}
+            className="text-blue-600 hover:underline font-medium"
+          >
+            ← Voltar
+          </Link>
+        </div>
+      </section>
     );
-
-    if (confirmed) {
-      setCategories(JSON.parse(JSON.stringify(originalCategories))); // Restaura do backup
-      setIsEditMode(false);
-      showFlash("Edição cancelada", "info");
-    }
-  };
-
-  // Salvar informações básicas do torneio
-  const handleSaveTournamentInfo = async (
-    name: string,
-    startDate: string,
-    endDate?: string
-  ) => {
-    if (!tournament?._id) return;
-
-    try {
-      await api.put(`/tournaments/${tournament._id}`, {
-        name,
-        startDate: dateToISO(startDate),
-        endDate: endDate ? dateToISO(endDate) : undefined,
-        date: dateToISO(startDate), // Compatibilidade
-      });
-
-      showFlash("Informações atualizadas com sucesso!", "success");
-      await fetchTournamentById(tournament._id);
-      setIsEditingInfo(false);
-    } catch (err: any) {
-      showFlash(
-        err.response?.data?.message || "Erro ao atualizar informações",
-        "error"
-      );
-    }
-  };
-
-  // Salvar categorias no backend
-  const handleSaveCategories = async () => {
-    if (!tournament?._id) return;
-
-    try {
-      await api.put(`/tournaments/${tournament._id}`, {
-        categories: categories,
-      });
-
-      showFlash("Categorias salvas com sucesso!", "success");
-      await fetchTournamentById(tournament._id);
-      setOriginalCategories(JSON.parse(JSON.stringify(categories)));
-      setIsEditMode(false);
-    } catch (err: any) {
-      showFlash(
-        err.response?.data?.message || "Erro ao salvar categorias!",
-        "error"
-      );
-    }
-  };
-
-  // Realizar sorteio de uma categoria
-  const handleDrawCategory = async (categoryId: string) => {
-    if (!tournament?._id) return;
-
-    const category = categories.find((c) => c.id === categoryId);
-    if (!category) return;
-
-    if (category.pairs.length < 4) {
-      showFlash(
-        "É necessário pelo menos 4 duplas para realizar o sorteio",
-        "warning"
-      );
-      return;
-    }
-
-    try {
-      const response = await api.post(
-        `/tournaments/${tournament._id}/draw-category`,
-        {
-          categoryId,
-        }
-      );
-
-      showFlash(
-        `Sorteio da categoria ${category.name} realizado com sucesso!`,
-        "success"
-      );
-
-      // Atualiza as categorias com os grupos gerados
-      const updatedCategories = categories.map((cat) => {
-        if (cat.id === categoryId) {
-          return {
-            ...cat,
-            groups: response.data.groups,
-            isDrawn: true,
-          };
-        }
-        return cat;
-      });
-
-      setCategories(updatedCategories);
-      await fetchTournamentById(tournament._id);
-    } catch (err: any) {
-      showFlash(
-        err.response?.data?.message || "Erro ao realizar sorteio",
-        "error"
-      );
-    }
-  };
-
-  // Atualiza placares modificados manualmente ao clicar no botão
-  const handleScoreUpdate = async () => {
-    if (!tournament?._id) return;
-
-    // Verifica autenticação
-    if (!isLoggedIn) {
-      showFlash(
-        "Você precisa estar logado para atualizar os placares",
-        "warning"
-      );
-      return;
-    }
-
-    // Verifica propriedade
-    if (!isOwner) {
-      showFlash(
-        "Apenas o criador do torneio pode atualizar os placares",
-        "error"
-      );
-      return;
-    }
-
-    const updates = [];
-
-    for (const [key, value] of Object.entries(scores)) {
-      const original = originalScores[key];
-      const [catIndex, gi, mi] = key.split("-").map(Number);
-
-      if (original?.team1 !== value.team1 || original?.team2 !== value.team2) {
-        const { team1, team2 } = value;
-        if (team1 !== "" && team2 !== "") {
-          updates.push({
-            categoryIndex: catIndex,
-            groupIndex: gi,
-            matchIndex: mi,
-            team1Score: Number(team1),
-            team2Score: Number(team2),
-          });
-        }
-      }
-    }
-
-    if (updates.length === 0) {
-      showFlash("Nenhuma alteração encontrada!", "info");
-      return;
-    }
-
-    try {
-      await api.put(`/tournaments/${tournament._id}/scores`, { updates });
-      await fetchTournamentById(tournament._id!);
-      showFlash("Placares atualizados com sucesso!", "success");
-
-      // Atualiza o snapshot original
-      setOriginalScores(scores);
-    } catch (err: any) {
-      console.error("Erro ao atualizar placares:", err);
-      showFlash(
-        err.response?.data?.message || "Erro ao atualizar placares",
-        "error"
-      );
-    }
-  };
-
-  if (loading) return <p className="p-6">Carregando...</p>;
-  if (!tournament) return <p className="p-6">Torneio não encontrado</p>;
+  }
 
   return (
     <section>
@@ -318,9 +96,9 @@ const TournamentDetails = () => {
           isEditingInfo={isEditingInfo}
           setIsEditMode={setIsEditMode}
           setIsEditingInfo={setIsEditingInfo}
-          onSave={handleSaveCategories}
-          onClean={handleCleanCategories}
-          onCancel={handleCancelEdit}
+          onSave={() => handleSaveCategories(categories)}
+          onClean={() => handleCleanCategories(resetCategories)}
+          onCancel={() => handleCancelEdit(resetCategories)}
         />
 
         {/* Editor de informações básicas */}
@@ -343,7 +121,9 @@ const TournamentDetails = () => {
           <CategoryManager
             categories={categories}
             onCategoriesChange={setCategories}
-            onDrawCategory={handleDrawCategory}
+            onDrawCategory={(categoryId) =>
+              handleDrawCategory(categoryId, categories, setCategories)
+            }
           />
         ) : (
           <CategoryList
@@ -351,14 +131,22 @@ const TournamentDetails = () => {
             canEdit={canEdit}
             scores={scores}
             onInputChange={setScores}
-            onUpdateScores={handleScoreUpdate}
+            onUpdateScores={() =>
+              handleScoreUpdate(
+                scores,
+                originalScores,
+                setOriginalScores,
+                isLoggedIn,
+                isOwner
+              )
+            }
           />
         )}
 
         {/* Navegação */}
         <div className="mt-8 mb-8 flex justify-between w-full max-w-6xl">
           <Link
-            to="/my-tournaments"
+            to={isLoggedIn ? "/my-tournaments" : "/"}
             className="text-blue-600 hover:underline font-medium"
           >
             ← Voltar
